+++ docs/examples/UserGuide/functions_pass_to_DB.jl
@@ -85,3 +85,22 @@ end
     @aside @show_query _
     @collect
 end
+
+# ## Interpolating queries
+# To use a prior, uncollected TidierDB query in other TidierDB macros, interpolate the needed query without showing or collecting it 
+ok = @chain t(df_mem) @summarize(mean = mean(value));
+# the mean value represented in SQL from the above is 3
+@eval @chain t(df_mem) begin 
+    @filter( value > $ok) 
+    @collect 
+end
+
+@eval @chain t(df_mem) begin 
+    @mutate(value2 =  value + $ok) 
+    @collect 
+end
+
+@eval @chain t(df_mem) begin 
+    @summarize(value =  mean(value) * $ok) 
+    @collect 
+end