+++ ext/LibPQExt.jl
@@ -42,7 +42,7 @@ function TidierDB.show_tables(con::LibPQ.Connection)
 end
 
 function TidierDB.final_compute(sqlquery::SQLQuery, ::Type{<:postgres}, sql_cr_or_relace::String=nothing)
-    final_query = finalize_query(sqlquery)
+    final_query = TidierDB.finalize_query(sqlquery)
     final_query = sql_cr_or_relace * final_query
     return LibPQ.execute(sq.db, final_query)
 end