+++ NEWS.md
@@ -1,4 +1,11 @@
 # TidierDB.jl updates
+## v0.5.1 - 2024-11-08
+- support for [reusing TidierDB queries](https://tidierorg.github.io/TidierDB.jl/latest/examples/generated/UserGuide/functions_pass_to_DB/#interpolating-queries) inside other macros, including `@mutate`, `@filter`, `@summarize`
+- adds `@union_all` to bind all rows not just distinct rows as with `@union`
+- joining syntax now supports `(table1, table2, col_name)` when joining columns have shared name
+- `if_else` now has optional final argument for handling missing values to match TidierData
+
+# TidierDB.jl updates
 ## v0.5.0 - 2024-10-15
 Breaking Changes: 
 - All join syntax now matches TidierData's `(table1, table2, t1_col = t2_col)`